+++ packages/terra-search-field/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Correctly applied mismatch tolerance.
+
 ## 3.72.0 - (September 28, 2021)
 
 * Changed